+++ Paramount.Betterclassifieds.Presentation/Services/IRobotVerifier.cs
@@ -4,6 +4,7 @@ namespace Paramount.Betterclassifieds.Presentation.Services
 {
     public interface IRobotVerifier
     {
+        bool IsValid(string siteSectionSecret, string googleCaptchaResult);
         bool IsValid(string siteSectionSecret, HttpRequestBase httpRequest);
     }
 }